MSD upgrade

+++ src/com/jjoe64/graphview/GraphViewDataInterface.java
@@ -20,11 +20,10 @@
 package com.jjoe64.graphview;
 
 /**
- * the base interface for the graphview data.
- * you can use your own data models, when they implement
- * this interface.
+ * the base interface for the graphview data. you can use your own data models, when they implement this interface.
  */
 public interface GraphViewDataInterface {
-	public double getX();
+	public long getX();
+
 	public double getY();
 }
